fixed dependencies and web socket bugs and minor typos

+++ src/main/java/com/ansh/authconnectionsexample/connectionpractice/dto/UserProfileDto.java
@@ -2,6 +2,7 @@ package com.ansh.authconnectionsexample.connectionpractice.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Set;
 
@@ -14,7 +15,7 @@ public class UserProfileDto {
     private Long id;
     private String username;
     private String experience;
-    private Double overallRating;
+    private BigDecimal overallRating;
     private Set<String> skill;
     private List<ReviewDto> reviewsReceived;
 }
